- Add tests for createUser() method (success, username exists, email exists, DB error) (#15)
* - Add tests for createUser() method (success, username exists, email exists, DB error)
- Add tests for getUserLoginByUsername() (success, not found, DB error)
- Add tests for getUserLoginByEmail() (success, not found, DB error)
- Refactor: use Mono.defer() in createUser() for lazy evaluation
- Refactor: remove redundant onErrorResume() from login methods

* - Add unit tests for UserManagementService (13 tests total)
- Tests cover createUser, getUserLoginByUsername, getUserLoginByEmail, updateLatestAccess
- Add Mockito verify() calls to ensure repository interactions
- Add security documentation for timing attack considerations
- Refactor: use Mono.defer() for proper lazy evaluation
- Refactor: add logging to updateLatestAccess for observability
- Fix: remove redundant onErrorResume() from login methods"

+++ MyData-service/src/main/java/Ecommerce/Reactive/MyData_service/service/CartService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.logging.Logger;
 
@@ -125,7 +126,7 @@ public class CartService {
                 .map(cart -> {
                     cart.setUserUuid(userUuid);
                     cart.setTotalProducts(0);
-                    cart.setTotalPrice(0.0);
+                    cart.setTotalPrice(BigDecimal.valueOf(0.0));
                     cart.setStatus(CartStatus.ACTIVE);
                     cart.setCreatedAt(LocalDateTime.now());
                     return cart;
